Fix for src\uqm\outfit.c(569) : warning C4146: unary minus operator applied to unsigned type, result still unsigned
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@3755 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/uqm/outfit.c
@@ -566,7 +566,7 @@ ChangeFuelQuantity (void)
 	{
 		const int maxFit = GetFuelTankCapacity () - GLOBAL_SIS (FuelOnBoard);
 		const int maxAfford = GLOBAL_SIS (ResUnits) / GLOBAL (FuelCost);
-		const int minFit = -GLOBAL_SIS (FuelOnBoard);
+		const int minFit = - (int) GLOBAL_SIS (FuelOnBoard);
 
 		if (incr > maxFit)
 			incr = maxFit; // All we can hold.
